requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.12.0
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
Flask==2.3.3
Flask-Login==0.6.2
Flask-Mail==0.9.1
flask-marshmallow==0.15.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.0.5
frozenlist==1.4.0
geoip2==4.7.0
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
maxminddb==2.4.0
multidict==6.0.4
packaging==23.1
psycopg2==2.9.7
psycopg2-binary==2.9.7
python-decouple==3.8
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.7
yarl==1.9.2
